KAFKA-18309: Change log.segment.bytes config type from INT to LONG
Change the log.segment.bytes (and its topic-level synonym segment.bytes)
config type from INT to LONG to allow segment sizes larger than ~2GB
(Integer.MAX_VALUE).

The INT type limited the maximum configurable segment size to
2,147,483,647 bytes (~1.99 GB). This change lifts that restriction by
using LONG type throughout the config definition, storage, and
propagation layers.

Changes:
- LogConfig: Change DEFAULT_SEGMENT_BYTES to long, config definitions
  from INT to LONG, segmentSize field/accessor from int to long
- AbstractKafkaConfig: logSegmentBytes() returns Long via getLong()
- RollParams: maxSegmentBytes field from int to long
- Cleaner: groupSegmentsBySize() maxSize param from int to long
- initFileSize() safely caps at Integer.MAX_VALUE for FileRecords
  compatibility

Backward/forward compatibility:
- Backward compatible: existing INT values (stored as strings) parse
  correctly as LONG
- Forward compatible: values <= Integer.MAX_VALUE work on older brokers;
  values > Integer.MAX_VALUE require the new version
- No wire protocol impact (broker/topic config only)

+++ core/src/test/scala/unit/kafka/log/LogTestUtils.scala
@@ -60,7 +60,7 @@ object LogTestUtils {
   }
 
   def createLogConfig(segmentMs: Long = LogConfig.DEFAULT_SEGMENT_MS,
-                      segmentBytes: Int = LogConfig.DEFAULT_SEGMENT_BYTES,
+                      segmentBytes: Long = LogConfig.DEFAULT_SEGMENT_BYTES,
                       retentionMs: Long = LogConfig.DEFAULT_RETENTION_MS,
                       localRetentionMs: Long = LogConfig.DEFAULT_LOCAL_RETENTION_MS,
                       retentionBytes: Long = ServerLogConfigs.LOG_RETENTION_BYTES_DEFAULT,
@@ -76,7 +76,7 @@ object LogTestUtils {
                       remoteLogDeleteOnDisable: Boolean = DEFAULT_REMOTE_LOG_DELETE_ON_DISABLE_CONFIG): LogConfig = {
     val logProps = new Properties()
     logProps.put(TopicConfig.SEGMENT_MS_CONFIG, segmentMs: java.lang.Long)
-    logProps.put(LogConfig.INTERNAL_SEGMENT_BYTES_CONFIG, segmentBytes: Integer)
+    logProps.put(LogConfig.INTERNAL_SEGMENT_BYTES_CONFIG, segmentBytes: java.lang.Long)
     logProps.put(TopicConfig.RETENTION_MS_CONFIG, retentionMs: java.lang.Long)
     logProps.put(TopicConfig.LOCAL_LOG_RETENTION_MS_CONFIG, localRetentionMs: java.lang.Long)
     logProps.put(TopicConfig.RETENTION_BYTES_CONFIG, retentionBytes: java.lang.Long)
